% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{add_cursor_coords_control}
\alias{add_cursor_coords_control}
\title{Add a cursor coordinates control to the map.}
\usage{
add_cursor_coords_control(
  map,
  position = "bottom-left",
  long_label = "Lng",
  lat_label = "Lat"
)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{position}{The position of the cursor coordinates control on the map.
Default is \code{"bottom-left"}.
Options include "top-left", "top-right", "bottom-left", "bottom-right".}

\item{long_label}{The label for the longitude coordinate. Default is \code{"Lng"}.}

\item{lat_label}{The label for the latitude coordinate. Default is \code{"Lat"}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       The map or map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Add a cursor coordinates control to the map.
}
