% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{add_custom_control}
\alias{add_custom_control}
\title{Add a custom HTML control to the map.}
\usage{
add_custom_control(map, id, html, position = "top-right")
}
\arguments{
\item{map}{The map or map proxy object.}

\item{id}{The ID for the custom control.}

\item{html}{The HTML content to add as a control.}

\item{position}{The position of the control on the map. Default is \code{"top-right"}.
Options include "top-left", "top-right", "bottom-left", "bottom-right".}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       The map or map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Add a custom HTML control to the map.
}
