% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_control.R
\name{add_draw_control}
\alias{add_draw_control}
\title{Add a draw control to the map}
\usage{
add_draw_control(
  map,
  id = "draw_control",
  position = "top-right",
  modes = c("polygon"),
  active_colour = "#04AAC1",
  inactive_colour = "#04AAC1",
  mode_labels = list(),
  features = NULL,
  panel_id = NULL,
  section_title = NULL,
  group_id = NULL
)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{id}{The ID for the draw control.}

\item{position}{The position of the draw control on the map. Default is \code{"top-right"}.
Options are "top-left", "top-right", "bottom-left", "bottom-right".}

\item{modes}{A vector of modes to enable in the draw control. Default is \code{c("polygon")}.
Options include "polygon", "delete", "line", and "point".}

\item{active_colour}{The colour for the drawn shapes. Default is \code{"#04AAC1"}.}

\item{inactive_colour}{The colour for the inactive shapes. Default is \code{"#04AAC1"}.}

\item{mode_labels}{A named list of labels for each mode.
For example, \code{list(polygon = "Draw Polygon", delete = "Delete Shape")}.}

\item{features}{Optional. An \code{sf} object or GeoJSON string of pre-drawn shapes to
populate the draw control with on initialisation.}

\item{panel_id}{ID of control panel to add to (optional).}

\item{section_title}{Section title when added to a control panel.}

\item{group_id}{Optional group ID for grouping controls within a panel.}
}
\value{
The map or map proxy object for chaining.
}
\description{
The draw control allows users to draw shapes (polygons, lines, points) on the map.
The drawn shapes can be styled and managed through the control options.
Information about the drawn shapes can be retrieved in Shiny using the
\code{input$map_shape_created} (where \code{map} is the ID of the map) reactive value.
}
\details{
For a more in-depth example see the
\href{https://epi-interactive-ltd.github.io/toro/articles/examples/controls/draw-control.html}{Draw control}
article.
}
\examples{
map() |>
 add_draw_control()
}
\seealso{
\code{\link[=get_drawn_shape]{get_drawn_shape()}} to retrieve the drawn shape as an \code{sf} object in Shiny.
}
