% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_fill_layer}
\alias{add_fill_layer}
\title{Add a fill layer to a map or map proxy}
\usage{
add_fill_layer(map, ...)
}
\arguments{
\item{map}{The map object or map proxy to which the layer will be added}

\item{...}{Additional arguments to include in the layer definition.
\itemize{
\item clusterOptions: A list of options for clustering, if \code{can_cluster} is \code{TRUE}.
See the \href{https://epi-interactive-ltd.github.io/toro/articles/layers.html}{cluster vignette}
for details on available options
}}
}
\value{
The updated map object with the fill layer added
}
\description{
Add a fill layer to a map or map proxy
}
\examples{
\dontrun{
# Load libraries
library(toro)
library(spData)
library(sf)

nz_data <- spData::nz |>
  rename(geometry = geom) |>
  sf::st_transform(4326)

map() |>
 add_fill_layer(
   id = "nz_regions",
   source = nz_data,
   hover_column = "Name"
 )

map() |>
 add_fill_layer(
   id = "nz_regions",
   source = nz_data,
   hover_column = "Name",
   paint = get_paint_options(
     "fill",
     options = list(
       colour = "#a3b18a",
       opacity = 0.3,
       outline_colour = "#588157"
     )
   )
 )
}
}
