% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{add_image}
\alias{add_image}
\title{Add an image source to the map.}
\usage{
add_image(map, image_id, image_url)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{image_id}{The ID of the image source.}

\item{image_url}{The URL of the image to add.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       The map or map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Add an image source to the map.
}
\examples{
if (interactive()) {
 map() |>
   add_image(
     image_id = "leaf-icon",
     image_url = "https://upload.wikimedia.org/wikipedia/en/thumb/0/02/Leaf_icon.png/600px-Leaf_icon.png"
   ) |>
   add_symbol_layer(
     id = "text_layer",
     source = sf::st_as_sf(quakes, coords = c("long", "lat"), crs = 4326),
     layout = toro::get_layout_options(
       "symbol",
       options = list(
         icon_image = "leaf-icon",
         icon_size = 0.1
       )
     )
   )
}
}
