% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_layer}
\alias{add_layer}
\title{Utilities for the map related to layers.
Add a layer to a map or map proxy.}
\usage{
add_layer(
  map,
  id,
  type = "fill",
  source,
  paint = NULL,
  layout = NULL,
  popup_column = NULL,
  hover_column = NULL,
  can_cluster = FALSE,
  under_id = NULL,
  filter = NULL,
  ...
)
}
\arguments{
\item{map}{The map object or map proxy to which the layer will be added.}

\item{id}{A unique identifier for the layer.}

\item{type}{The type of layer to add (e.g., "fill", "circle", "line").
Default is "fill".}

\item{source}{The data source for the layer, if not a GeoJSON, it will be converted.}

\item{paint}{A list of paint options for styling the layer. See \code{get_paint_options()}
for defaults and options.}

\item{layout}{A list of layout options for the layer. See \code{get_layout_options()} for defaults
and options.}

\item{popup_column}{The column name to use for popups. Default is \code{NULL}.}

\item{hover_column}{The column name to use for hover effects. Default is \code{NULL}.}

\item{can_cluster}{Whether the layer can be clustered. Default is \code{FALSE}.}

\item{under_id}{The ID of an layer already on the map to place this layer under.
Default is \code{NULL}.}

\item{filter}{A filter expression to apply to the layer. Default is \code{NULL}. See
\code{get_layer_filter()} for more details on how to create filter expressions.}

\item{...}{Additional arguments to include in the layer definition.
\itemize{
\item clusterOptions: A list of options for clustering, if \code{can_cluster} is \code{TRUE}.
See the \href{https://epi-interactive-ltd.github.io/toro/articles/layers.html}{cluster vignette}
for details on available options.
}}
}
\value{
The updated map object with the new layer added.
}
\description{
Utilities for the map related to layers.
Add a layer to a map or map proxy.
}
\note{
If \code{source} is not a string referring to an existing source, it will be converted to a
GeoJSON source and added to the map automatically. The Id for the source is generated by
appending the layer Id to \verb{source-}. For example, if you add a layer with \code{id = "my_layer"}
and \code{source} is a sf object, the source is added to the map with the ID \code{"source-my_layer"}.
}
