% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_line_layer}
\alias{add_line_layer}
\title{Add a line layer to a map or map proxy.}
\usage{
add_line_layer(map, ...)
}
\arguments{
\item{map}{The map object or map proxy to which the layer will be added.}

\item{...}{Additional arguments to include in the layer definition.
\itemize{
\item clusterOptions: A list of options for clustering, if \code{can_cluster} is \code{TRUE}.
See the \href{https://epi-interactive-ltd.github.io/toro/articles/layers.html}{cluster vignette}
for details on available options.
}}
}
\value{
The updated map object with the line layer added.
}
\description{
Add a line layer to a map or map proxy.
}
\examples{
\dontrun{
# Load libraries
library(toro)
library(spData)
library(sf)

seine_data <- spData::nz_height |>
  sf::st_transform(4326)
seine_data$colour <- rainbow(nrow(seine_data))

map() |>
 set_bounds(bounds = seine_data) |>
 add_circle_layer(
   id = "seine_lines",
   source = seine_data,
   hover_column = "name"
 )

base_map |>
 set_bounds(bounds = seine_data) |>
 add_circle_layer(
   id = "seine_lines",
   source = seine_data,
   hover_column = "name",
   paint = get_paint_options(
     "circle",
     options = list(colour = get_column("colour"))
   )
 )
}
}
