% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_control.R
\name{add_paint_control}
\alias{add_paint_control}
\title{Add a paint control to the map}
\usage{
add_paint_control(
  map,
  id,
  layer_id,
  paint_options,
  label = NULL,
  default = NULL,
  position = "top-right",
  input_type = "select"
)
}
\arguments{
\item{map}{A toro map or mapProxy object.}

\item{id}{Unique identifier for the paint control.}

\item{layer_id}{The ID of the layer to control.}

\item{paint_options}{A list of paint options to control
(e.g., list(
list(id = "default", label = "Default", paint = list('fill-color' = 'red')),
list(id = "alternative", label = "Alternative", paint = list('fill-color' = 'blue'))
)).}

\item{label}{Optional label for the control.}

\item{default}{Optional default value for the control.}

\item{position}{Position of the control on the map (e.g., "top-right").}

\item{input_type}{Type of input control. Either "select" for a dropdown or "radio" for radio
buttons. Default is "select".}
}
\value{
The modified map object.
}
\description{
Allow for a layer to change paint properties based on user input.
}
\examples{
library(shiny)
library(dplyr)
library(sf)
library(toro)


ui <- fluidPage(
  mapOutput("map"),
  actionButton("remove_control", "Remove paint control"),
  actionButton("add_control", "Add paint control")
)

nz_long <- c(
  172.2041,
  182.2494,
  168.1286,
  163.9383,
  171.7449
)
nz_lat <- c(
  -32.56960,
  -38.85522,
  -48.49620,
  -46.43999,
  -34.62533
)

nz_line <- sf::st_sf(
  id = 1,
  geometry = sf::st_sfc(sf::st_linestring(cbind(nz_long, nz_lat)), crs = 4326)
)

nz_points <- sf::st_cast(nz_line, "POINT") |>
  mutate(id = row_number())

nz_points$date <- seq.Date(
  from = as.Date("2026-01-01"),
  to = as.Date("2026-01-05"),
  by = "day"
)
nz_points$lat <- nz_lat
nz_points$long <- nz_long
nz_points$tici <- seq_len(nrow(nz_points))
nz_points$group <- c("A", "B", "A", "C", "D")
nz_points$colour <- c("red", "pink", "green", "orange", "blue")
nz_points$bool <- c(FALSE, FALSE, FALSE, TRUE, TRUE)

paint_options <- list(
  list(
    id = "default",
    label = "Default",
    paint = get_paint_options(
      "circle",
      list(
        colour = "#FFCD00",
        circle_radius = 5,
        outline_colour = "black"
      )
    )
  ),
  list(
    id = "colour",
    label = "Colour",
    paint = get_paint_options(
      "circle",
      list(
        colour = get_column("colour"),
        circle_radius = 6,
        outline_colour = "white"
      )
    )
  ),
  list(
    id = "bool",
    label = "Bool",
    paint = get_paint_options(
      "circle",
      list(
        colour = get_column_boolean("bool", "white", "red"),
        circle_radius = 6,
        outline_colour = "white"
      )
    )
  ),
  list(
    id = "steps",
    label = "Steps",
    paint = get_paint_options(
      "circle",
      list(
        colour = get_column_steps(
          "long",
          c(171, 172, 178),
          c("red", "orange", "yellow", "green")
        ),
        circle_radius = 6,
        outline_colour = "white"
      )
    )
  ),
  list(
    id = "group",
    label = "Group",
    paint = get_paint_options(
      "circle",
      list(
        colour = get_column_group(
          "group",
          c("A" = "red", "B" = "yellow"),
          "white"
        ),
        circle_radius = 6,
        outline_colour = "white",
        opacity = get_column_group(
          "group",
          stats::setNames(c(0.9, 0.7), c("A", "B")),
          0.3
        )
      )
    )
  )
)

server <- function(input, output, session) {
  # Render UI -------------------------------------------------------

  output$map <- toro::renderMap({
    toro::map() |>
      set_zoom(3) |>
      add_source("my_source", nz_points) |>
      add_circle_layer("circle_layer", "my_source") |>
      add_paint_control(
        "my_paint_control",
        "circle_layer",
        paint_options,
        label = "Index",
        default = "steps"
      )
  })

  # Observers -------------------------------------------------------------

  observe({
    req(input$map_loaded)
    remove_paint_control(mapProxy("map"), "my_paint_control")
  }) |>
    bindEvent(input$remove_control)

  observe({
    req(input$map_loaded)
    add_paint_control(
      mapProxy("map"),
      "my_paint_control",
      "circle_layer",
      paint_options,
      default = "bool"
    )
  }) |>
    bindEvent(input$add_control)
}

if (interactive()) {
  shinyApp(ui, server)
}
}
