% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{add_source}
\alias{add_source}
\title{Utilities for the map related to map sources.
Add a source to the map.}
\usage{
add_source(map, source_id, data, type = "geojson", cluster = FALSE, ...)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{source_id}{The ID for the source.}

\item{data}{The data for the source, typically in GeoJSON format.}

\item{type}{The type of the source. Default is \code{"geojson"}.
Other options include \code{"vector"} or \code{"raster"}.}

\item{cluster}{Whether to enable clustering for this source. Default is \code{FALSE}.}
}
\value{
The map or map proxy object for chaining.
}
\description{
Utilities for the map related to map sources.
Add a source to the map.
}
\examples{
\dontrun{
 map() |>
   add_source(
     source_id = "my_source",
     data = sf::st_as_sf(quakes, coords = c("long", "lat"), crs = 4326)
   ) |>
   add_circle_layer(id = "quakes", source = "my_source")
}
}
