% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{add_speed_control}
\alias{add_speed_control}
\title{Add a speed control to the map or control panel}
\usage{
add_speed_control(
  map,
  values = c(0.5, 1, 2),
  labels = c("Slow", "Normal", "Fast"),
  default_index = 2,
  position = "top-right",
  panel_id = NULL,
  section_title = NULL,
  group_id = NULL
)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{values}{Vector of speed multiplier values. Default is c(0.5, 1, 2).}

\item{labels}{Vector of labels for each speed value. Default is c("Slow", "Normal", "Fast").}

\item{default_index}{Index of the default speed (1-based). Default is 2.}

\item{position}{Position on the map if not using a control panel. Default is "top-right".}

\item{panel_id}{ID of control panel to add to (optional).}

\item{section_title}{Section title when added to a control panel.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         The map or map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Add a speed control to the map or control panel
}
