% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{add_tiles_from_map_server}
\alias{add_tiles_from_map_server}
\title{Add an Image Server source to a toro map}
\usage{
add_tiles_from_map_server(map, url, tile_id, as_image_layer = FALSE, ...)
}
\arguments{
\item{map}{A toro map object or a map proxy object.}

\item{url}{The URL of the ArcGIS Image Server.}

\item{...}{Additional parameters for the Image Server source.}

\item{source_id}{A unique identifier for the source.}
}
\value{
The updated map object.
}
\description{
Add an Image Server source to a toro map
}
