% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_controls.R
\name{add_zoom_control}
\alias{add_zoom_control}
\title{Functions related to zoom controls.
Add a zoom control to the map.}
\usage{
add_zoom_control(
  map,
  position = "top-right",
  control_options = list(),
  panel_id = NULL,
  section_title = NULL,
  group_id = NULL
)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{position}{The position of the zoom control on the map. Default is \code{"top-right"}.}

\item{control_options}{Additional options for the zoom control. Default is an empty list.}

\item{panel_id}{ID of control panel to add to (optional).}

\item{section_title}{Section title when added to a control panel.}
}
\value{
The map proxy object for chaining.
}
\description{
Functions related to zoom controls.
Add a zoom control to the map.
}
\note{
See \href{https://maplibre.org/maplibre-gl-js/docs/API/type-aliases/NavigationControlOptions/}{Maplibre NavigationControl docs}
for more information on available options.
}
\examples{
\dontrun{
add_zoom_control(map())

# Inside a control panel
map() |>
 add_control_panel(panel_id = "my_panel", title = "View Controls") |>
 add_zoom_control(panel_id = "my_panel")
}

}
