% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArcGisDependencies.R
\name{addArcGisDependencies}
\alias{addArcGisDependencies}
\title{Add ArcGis dependencies to a Shiny app}
\usage{
addArcGisDependencies()
}
\value{
A list of HTML tags to be included in the Shiny app's head.
}
\description{
Include this inside \code{tags$head} to add both the CSS and JS dependencies
for ArcGis.
}
