% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMaplibreDependencies.R
\name{addMaplibreDependencies}
\alias{addMaplibreDependencies}
\title{Add Maplibre dependencies to a Shiny app}
\usage{
addMaplibreDependencies()
}
\value{
A list of HTML tags to be included in the Shiny app's head.
}
\description{
Include this inside \code{tags$head} to add both the CSS and JS dependencies
for Maplibre.
}
