% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{download_map_image}
\alias{download_map_image}
\title{Utilities for the map related to exports.}
\usage{
download_map_image(proxy, filename, format = "png", width = 800, height = 600)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{filename}{The name of the file to save the image as.}

\item{format}{The format of the image. Default is "png".}

\item{width}{The width of the image in pixels. Default is 800.}

\item{height}{The height of the image in pixels. Default is 600.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    The map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Functions:
\itemize{
\item \code{download_map_image}: Download the map as an image.
Download the map as an image.
}
}
\note{
THIS IS A WIP !!!
}
\keyword{internal}
