% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_values.R
\name{get_column_group}
\alias{get_column_group}
\title{Get the values for a paint or layout option in a map layer based on a column value.}
\usage{
get_column_group(column_name, named_group_values, default_value = "#cccccc")
}
\arguments{
\item{column_name}{String representing the name of the column to be used.}

\item{named_group_values}{Vector of value strings named by the group values.
The names of the vector should match the group values in the column.}

\item{default_value}{String for the default value to use if no match is found.
Default is "#cccccc".}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{              List containing the paint or layout option to be set.
}\if{html}{\out{</div>}}
}
\description{
Allows the data to be styled by the group option in the column.
}
\note{
If using numbers as the group values, then you need to use \code{setNames} rather
than a named vector, as the names of the vector will be coerced to strings.
}
\examples{
get_column_group("group", c("A" = "red", "B" = "blue"), "grey")
get_column_group("opacity", setNames(c(0.3, 0.5), c("A", "B")), 0.6)

}
