% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layer_options.R
\name{get_layer_filter}
\alias{get_layer_filter}
\title{Get a filter for a layer}
\usage{
get_layer_filter(filter_str)
}
\arguments{
\item{filter_str}{A string or vector of strings representing the filter conditions.}
}
\value{
A list where the first element is "all" if multiple filters are provided,
or a single filter condition.
}
\description{
Parse a filter string into a list of filters that the map can use.
}
\examples{
# Filter to only show rows where the "layer_id" column is equal to "forests"
get_layer_filter("layer_id == forests")

# Filter to show rows where the "layer_id" column is equal to "sites" and the "project_status"
# column is equal to "Confirmed"
get_layer_filter(c("layer_id == sites", "project_status == Confirmed"))
}
