% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{get_tile_options}
\alias{get_tile_options}
\title{Get available tile layer options.}
\usage{
get_tile_options()
}
\value{
A character vector of available tile layer options.
}
\description{
Get available tile layer options.
}
