% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClickedFeature.R
\name{getClickedFeature}
\alias{getClickedFeature}
\title{Get the sf data frame of a clicked feature from the map widget.}
\usage{
getClickedFeature(clicked_feature_input)
}
\arguments{
\item{clicked_feature_input}{A list representing the drawn shape.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{                 A `sf` object representing the clicked feature, or `NULL`.
}\if{html}{\out{</div>}}
}
\description{
The click input is a list containing the \code{layerId}, \code{properties}, \code{geometry}, and
\code{time}. Turn this into an \code{sf} object.
}
\note{
\code{time} is not used in this function, but it is included in the input so that
the same feature can be clicked multiple times and the changed time means that the
input will be updated.
}
