% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumn.R
\name{getColumn}
\alias{getColumn}
\title{Get a column from a dataset to use as a paint or layout option in a map layer.}
\usage{
getColumn(column_name)
}
\arguments{
\item{column_name}{String representing the name of the column to be used.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      List containing the paint or layout option to be set.
}\if{html}{\out{</div>}}
}
\description{
Allows the column value to be used for styling features in a map layer.
}
\examples{
getColumn("opacity")
getColumn("icon")
# Use in a paint property: list("circle-color" = getColumn("color"))

}
