% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumnBooleanValues.R
\name{getColumnBooleanValues}
\alias{getColumnBooleanValues}
\title{Get the value for a paint or layout option in a map layer based on a column boolean value.}
\usage{
getColumnBooleanValues(column_name, true_value, false_value)
}
\arguments{
\item{column_name}{String representing the name of the column to be used.}

\item{true_value}{Value to use when the column value is \code{TRUE}.}

\item{false_value}{Value to use when the column value is \code{FALSE}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      List containing the paint or layout option to be set.
}\if{html}{\out{</div>}}
}
\description{
Allows the data to be styled by the group option in the column.
}
\examples{
getColumnBooleanValues("group", c("A" = "red", "B" = "blue"), "grey")

}
