% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumnGroupColours.R
\name{getColumnGroupColours}
\alias{getColumnGroupColours}
\title{Get the colours for a paint or layout option in a map layer based on a column value.}
\usage{
getColumnGroupColours(
  column_name,
  named_group_colours,
  default_colour = "#cccccc"
)
}
\arguments{
\item{column_name}{String representing the name of the column to be used.}

\item{named_group_colours}{Vector of colour strings named by the group values.
The names of the vector should match the group values in the column.}

\item{default_colour}{String for the default colour to use if no match is found.
Default is "#cccccc".}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{              List containing the paint or layout option to be set.
}\if{html}{\out{</div>}}
}
\description{
Allows the data to be styled by the group option in the column.
}
\examples{
getColumnGroupColours("group", c("A" = "red", "B" = "blue"), "grey")

}
