% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumnStepColours.R
\name{getColumnStepColours}
\alias{getColumnStepColours}
\title{Get the colours for a column in a map layer based on step breaks.}
\usage{
getColumnStepColours(column_name, breaks, colours)
}
\arguments{
\item{column_name}{String representing the name of the column to be used.}

\item{breaks}{Numeric vector of thresholds (must be sorted ascending).}

\item{colours}{Vector of colours, length = length(breaks) + 1.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{          List containing the paint or layout option to be set.
}\if{html}{\out{</div>}}
}
\description{
Allows the data to be styled by the step breaks in the column.
}
\examples{
getColumnStepColours("value", c(10, 20, 30), c("red", "orange", "yellow", "green", "blue"))

}
