% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrawnShape.R
\name{getDrawnShape}
\alias{getDrawnShape}
\title{Get the drawn shape from the map widget.}
\usage{
getDrawnShape(create_input_string)
}
\arguments{
\item{create_input_string}{A JSON string representing the drawn shape.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{               A `sf` object representing the drawn shape, or `NULL`.
}\if{html}{\out{</div>}}
}
\description{
Parses the JSON string returned by the map widget when a shape is drawn.
Ensures that the ID of the shape is included in the resulting \code{sf} object.
}
