% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLayoutOptions.R
\name{getLayoutOptions}
\alias{getLayoutOptions}
\title{Get layout options for a specific layer type}
\usage{
getLayoutOptions(layer_type, options = list())
}
\arguments{
\item{layer_type}{A string indicating the type of layer (e.g., "fill", "circle", "line").}

\item{options}{A list of additional options to customize the layout properties.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A list of layout options suitable for the specified layer type.
}\if{html}{\out{</div>}}
}
\description{
This function returns a list of layout options based on the layer type and any additional
options provided.
}
\examples{
getLayoutOptions("line", list(line_cap = "butt", line_join = "bevel"))
getLayoutOptions("symbol", list(icon_image = "yellow_pin", icon_size = 1.5))

}
\seealso{
\code{\link{getColumn}}, \code{\link{getColumnGroupColours}},
\code{\link{getColumnStepColours}}
}
