% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPaintOptions.R
\name{getPaintOptions}
\alias{getPaintOptions}
\title{Get paint options for a specific layer type}
\usage{
getPaintOptions(layer_type, options = list())
}
\arguments{
\item{layer_type}{A string indicating the type of layer (e.g., "fill", "circle", "line").}

\item{options}{A list of additional options to customize the paint properties.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A list of paint options suitable for the specified layer type.
}\if{html}{\out{</div>}}
}
\description{
This function returns a list of paint options based on the layer type and any
additional options provided.
}
\examples{
getPaintOptions("line", list(colour = "blue", opacity = 0.8, line_width = 2))
getPaintOptions("circle", list(colour = "red", circle_radius = 10, outline_colour = "black"))
# Use with getColumn for data-driven styling:
getPaintOptions("fill", list(colour = getColumn("color"), opacity = getColumn("opacity")))
getPaintOptions("fill", list(
   colour = getColumnGroupColours("group", c("A" = "green", "B" = "blue")
))
getPaintOptions("fill", list(
   opacity = getColumnStepColours("percent", c(25, 75), c("red", "orange", "yellow"))
))

}
\seealso{
\code{\link{getColumn}}, \code{\link{getColumnGroupColours}},
\code{\link{getColumnStepColours}}
}
