% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{image_to_data_uri}
\alias{image_to_data_uri}
\title{Helper function to convert local image to data URI}
\usage{
image_to_data_uri(image_path)
}
\arguments{
\item{image_path}{The file path to the local image.}
}
\value{
A data URI string representing the image.
}
\description{
Helper function to convert local image to data URI
}
\keyword{internal}
