% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layer_options.R
\name{layerFilter}
\alias{layerFilter}
\title{Get a filter for a layer}
\usage{
layerFilter(filter_str)
}
\arguments{
\item{filter_str}{A string or vector of strings representing the filter conditions.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A list where the first element is "all" if multiple filters are provided,
               or a single filter condition.
}\if{html}{\out{</div>}}
}
\description{
Parse a filter string into a list of filters that the map can use.
}
\examples{
layerFilter("layer_id == forests")
layerFilter(c("layer_id == sites", "project_status == Confirmed"))

}
