% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapProxy}
\alias{mapProxy}
\title{Create a proxy object for updating the map}
\usage{
mapProxy(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{The ID of the output element}

\item{session}{The Shiny session object (default is the current session)}
}
\value{
A proxy object for the map
}
\description{
Create a proxy object for updating the map
}
\examples{
\dontrun{
library(shiny)
library(toro)

ui <- fluidPage(
 tagList(
   mapOutput("map"),
   checkboxInput("has_zoom_controls", "Remove Zoom Controls", value = TRUE)
 )
)
server <- function(input, output, session) {
 output$map <- renderMap({
   map() |>
     add_zoom_control()
 })

 observe({
   req(input$map_loaded)
   if (input$has_zoom_controls == TRUE) {
     mapProxy("map") |>
       add_zoom_control()
   } else {
     mapProxy("map") |>
       remove_zoom_control()
   }
 }) |>
   bindEvent(input$has_zoom_controls)
}
}
}
