% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{pause_route}
\alias{pause_route}
\title{Pause a route animation on a toro map.}
\usage{
pause_route(map, route_id, settings = list())
}
\arguments{
\item{map}{A toro map proxy object.}

\item{route_id}{A unique identifier for the route.}

\item{settings}{A list of settings for pausing the animation.}
}
\value{
The updated map object.
}
\description{
Pause a route animation on a toro map.
}
