% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{play_route}
\alias{play_route}
\title{Play a route animation on a toro map.}
\usage{
play_route(map, route_id, settings = list())
}
\arguments{
\item{map}{A toro map object or a map proxy object.}

\item{route_id}{A unique identifier for the route.}

\item{settings}{A list of settings for the animation (e.g., speed, loop).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     The updated map object.
}\if{html}{\out{</div>}}
}
\description{
Play a route animation on a toro map.
}
