% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{remove_control_from_panel}
\alias{remove_control_from_panel}
\title{Remove a control from a control panel.}
\usage{
remove_control_from_panel(proxy, panel_id, control_id)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{panel_id}{The ID of the control panel.}

\item{control_id}{The ID of the control to remove from the panel.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       The map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Remove a control from a control panel.
}
