% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_panel.R
\name{remove_control_group}
\alias{remove_control_group}
\title{Remove a control group from a control panel}
\usage{
remove_control_group(proxy, panel_id, group_id)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{panel_id}{The ID of the control panel.}

\item{group_id}{The ID of the control group to remove.}
}
\value{
The map proxy object for chaining.
}
\description{
Remove a control group from a control panel
}
\examples{
if(interactive()){
library(shiny)
library(toro)

ui <- fluidPage(
 tagList(
   mapOutput("map"),
   actionButton("remove_group1", "Remove control group 1")
 )
)
server <- function(input, output, session) {
 output$map <- renderMap({
   map() |>
     add_control_panel(panel_id = "my_panel", title = "Map Settings") |>
     add_control_group(
       panel_id = "my_panel",
       group_id = "control_group1",
       group_title = "Control Group 1"
     ) |>
    add_control_group(
       panel_id = "my_panel",
       group_id = "control_group2",
       group_title = "Control Group 2"
     )
 })

 observe({
   req(input$map_loaded)
   mapProxy("map") |>
     remove_control_group(panel_id = "my_panel", group_id = "control_group1")
 }) |>
   bindEvent(input$remove_group1)
}
}
}
