% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{remove_route}
\alias{remove_route}
\title{Remove an animation route from a toro map.}
\usage{
remove_route(map, route_id, settings = list())
}
\arguments{
\item{map}{A toro map object or a map proxy object.}

\item{route_id}{A unique identifier for the route.}

\item{settings}{A list of settings for removing the route.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     The updated map object.
}\if{html}{\out{</div>}}
}
\description{
Remove an animation route from a toro map.
}
