% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_control.R
\name{remove_scale_control}
\alias{remove_scale_control}
\title{Remove a scale control from the map}
\usage{
remove_scale_control(proxy, id = "scale_control")
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{id}{Unique identifier of the scale control to remove.
Default is \code{"scale_control"}.}
}
\value{
The map proxy object for chaining.
}
\description{
Remove a scale control from the map
}
