% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_controls.R
\name{remove_tile_selector_control}
\alias{remove_tile_selector_control}
\title{Remove the tile selector control from the map}
\usage{
remove_tile_selector_control(proxy, panel_id = NULL)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}}

\item{panel_id}{Optional. If provided, removes the tile selector control from the specified
control panel. If NULL, removes the standalone tile selector control}
}
\value{
The map proxy object for chaining
}
\description{
Remove the tile selector control from the map
}
\examples{
\dontrun{
library(shiny)
library(toro)

all_tiles <- get_tile_options()

ui <- fluidPage(
 tagList(
   mapOutput("map"),
   actionButton("remove_control", "Remove tile selector control")
 )
)
server <- function(input, output, session) {
 output$map <- renderMap({
   map(loadedTiles = all_tiles) |>
     add_tile_selector_control(available_tiles = all_tiles)
 })

 observe({
   req(input$map_loaded)
   mapProxy("map") |>
     remove_tile_selector_control()
 }) |>
   bindEvent(input$remove_control)
}
}
}
