% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{save_map_html}
\alias{save_map_html}
\title{Save map as standalone HTML file.}
\usage{
save_map_html(map, filepath, title = "Toro Map", selfcontained = TRUE, ...)
}
\arguments{
\item{map}{A map object created by \code{map()}.}

\item{filepath}{The file path to save the HTML file (should end with .html).}

\item{title}{The title for the HTML page. Default is "Toro Map".}

\item{selfcontained}{Whether to create a self-contained HTML file. Default is TRUE.}

\item{...}{Additional arguments passed to htmlwidgets::saveWidget().}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    The file path of the saved HTML file (invisibly).
}\if{html}{\out{</div>}}
}
\description{
This function saves a map widget as a self-contained HTML file that can be
opened in any web browser.
}
\examples{
\dontrun{
# Create and save a map as HTML
my_map <- map() |>
  add_source("earthquakes", quakes_data) |>
  add_circle_layer("quake_circles", source = "earthquakes")

save_map_html(my_map, "earthquake_map.html", title = "Earthquake Data")
}
}
