% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{set_bounds}
\alias{set_bounds}
\title{Set the map bounds.}
\usage{
set_bounds(map, bounds, padding = 50, max_zoom = map$maxZoom)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{bounds}{One of two formats:
\itemize{
\item A list of two coordinate pairs: \code{list(list(lon1, lat1), list(lon2, lat2))}.
\item An \code{sf} object, which will be converted to a bounding box.
}}

\item{padding}{The padding around the bounds in pixels. Default is 50.}

\item{max_zoom}{The maximum zoom level to set. Default is the object's \code{maxZoom}.}
}
\value{
The map or map proxy object for chaining.
}
\description{
Set the map bounds.
}
