% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{set_tile_layer}
\alias{set_tile_layer}
\title{Set the tile layer for the map.}
\usage{
set_tile_layer(map, tiles)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{tiles}{A character vector of tile layer names.
Options include values returned from get_tile_options().}
}
\value{
The map or map proxy object for chaining.
}
\description{
Set the tile layer for the map.
}
\examples{
\dontrun{
library(shiny)
library(toro)

all_tiles <- get_tile_options()

ui <- fluidPage(
 tagList(
   mapOutput("map"),
   selectInput("tile_layer", "Select Tile Layer", choices = all_tiles)
 )
)
server <- function(input, output, session) {
 output$map <- renderMap({
   map(loadedTiles = all_tiles)
 })

 observe({
   mapProxy("map") |>
     set_tile_layer(tiles = input$tile_layer)
 }) |>
   bindEvent(input$tile_layer)
}
}
}
\seealso{
get_tile_options
}
