% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{show_layer}
\alias{show_layer}
\title{Show a previously hidden layer on the map}
\usage{
show_layer(proxy, layer_id)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}}

\item{layer_id}{The ID of the layer to show}
}
\value{
The map proxy object for chaining
}
\description{
Show a previously hidden layer on the map
}
\examples{
\dontrun{
library(shiny)
library(sf)
library(toro)

data(quakes)
quakes_data <- st_as_sf(quakes, coords = c("long", "lat"), crs = 4326)

ui <- fluidPage(
 tagList(
   mapOutput("map"),
   actionButton("show_layer", "Show Layer"),
   actionButton("hide_layer", "Hide Layer")
 )
)
server <- function(input, output, session) {
 output$map <- renderMap({
   map() |>
     add_circle_layer(
       id = "quakes",
       source = quakes_data
     )
 })

 observe({
   mapProxy("map") |>
     show_layer(layer_id = "quakes")
 }) |>
   bindEvent(input$show_layer)

 observe({
   mapProxy("map") |>
     hide_layer(layer_id = "quakes")
 }) |>
   bindEvent(input$hide_layer)
}
}
}
