% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{toggle_control}
\alias{toggle_control}
\title{Toggle the visibility of a control on the map}
\usage{
toggle_control(proxy, control_id, show = TRUE)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}}

\item{control_id}{The ID of the control to toggle}

\item{show}{Logical indicating whether to show or hide the control. Default is \code{TRUE}}
}
\value{
The map proxy object for chaining
}
\description{
Toggle the visibility of a control on the map
}
\examples{
\dontrun{
# In a Shiny app:
output$map <- renderMap({
 map() |>
   add_zoom_control() |>
   add_custom_control(
     id = "custom_control",
     html = "<p>I am a custom control</p>"
   )
})

# In an observer:
mapProxy("map") |>
 toggle_control("zoom_control", show = FALSE) |>
 toggle_control("custom_control", show = FALSE)
}
}
