% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_control.R
\name{toggle_scale_control}
\alias{toggle_scale_control}
\title{Show or hide a scale control on the map}
\usage{
toggle_scale_control(proxy, id = "scale_control", show = TRUE)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{id}{Unique identifier of the scale control to toggle.
Default is \code{"scale_control"}.}

\item{show}{Logical indicating whether to show or hide the control.
Default is \code{TRUE}.}
}
\value{
The map proxy object for chaining.
}
\description{
Show or hide a scale control on the map
}
\examples{
if (interactive()) {
library(shiny)
library(toro)

ui <- fluidPage(
  tagList(
    mapOutput("map"),
    checkboxInput("show_scale", "Show scale control", value = TRUE)
  )
)
server <- function(input, output, session) {
  output$map <- renderMap({
    map() |> add_scale_control()
  })

  observe({
    req(input$map_loaded)
    mapProxy("map") |>
      toggle_scale_control(show = input$show_scale)
  }) |>
    bindEvent(input$show_scale)
}
}
}
