% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{validate_bounds}
\alias{validate_bounds}
\title{Get the bounds in the correct format}
\usage{
validate_bounds(bounds)
}
\arguments{
\item{bounds}{A list of two coordinate pairs or an sf object.}
}
\value{
A list of two coordinate pairs in the format \code{list(list(lon1, lat1), list(lon2, lat2))}.
}
\description{
This function takes either a list of two coordinate pairs or an sf object and returns the bounds
in the correct format for the JS function.
}
\keyword{internal}
