% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_circle_layer}
\alias{add_circle_layer}
\title{Add a circle layer to a map or map proxy}
\usage{
add_circle_layer(map, ...)
}
\arguments{
\item{map}{The map object or map proxy to which the layer will be added}

\item{...}{Additional arguments to include in the layer definition.
\itemize{
\item clusterOptions: A list of options for clustering, if \code{can_cluster} is \code{TRUE}.
See the \href{https://epi-interactive-ltd.github.io/toro/articles/layers.html}{cluster vignette}
for details on available options
}}
}
\value{
The updated map object with the circle layer added
}
\description{
Add a circle layer to a map or map proxy
}
\examples{
\dontrun{
# Load libraries
library(toro)
library(spData)
library(sf)

nz_data <- spData::nz_height |>
  sf::st_transform(4326)

map() |>
 set_bounds(bounds = nz_data) |>
 add_circle_layer(
   id = "nz_elevation",
   source = nz_data,
   hover_column = "elevation"
 )

base_map |>
 set_bounds(bounds = nz_data) |>
 add_circle_layer(
   id = "nz_elevation",
   source = nz_data,
   hover_column = "elevation",
   paint = get_paint_options(
     "circle",
     options = list(
       colour = get_column_step_steps(
         "elevation",
         c(3000),
         c("grey", "black")
       )
     )
   )
 )
}
}
