% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_circle_layer}
\alias{add_circle_layer}
\title{Add a circle layer to a map or map proxy.}
\usage{
add_circle_layer(map, ...)
}
\arguments{
\item{map}{The map object or map proxy to which the layer will be added.}
}
\value{
The updated map object with the circle layer added.
}
\description{
Add a circle layer to a map or map proxy.
}
