% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{add_control_panel}
\alias{add_control_panel}
\title{Add a control panel to the map}
\usage{
add_control_panel(
  map,
  panel_id,
  title = NULL,
  position = "bottom-left",
  collapsible = FALSE,
  collapsed = FALSE,
  direction = "column",
  custom_controls = NULL
)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{panel_id}{Unique identifier for the control panel.}

\item{title}{Title for the control panel. If NULL, no title is shown.}

\item{position}{Position of the control panel on the map. Default is "bottom-left".
Options include "top-left", "top-right", "bottom-left", "bottom-right".}

\item{collapsible}{Whether the panel can be collapsed. Default is FALSE.}

\item{collapsed}{Initial collapsed state. Default is FALSE.}

\item{custom_controls}{List of custom controls to add initially. Each should be a list with
elements: html, id (optional), title (optional).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         The map or map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Creates a flexible control panel that can contain multiple controls.
}
