% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{add_custom_control}
\alias{add_custom_control}
\title{Add a custom HTML control to the map.}
\usage{
add_custom_control(
  map,
  id,
  html,
  position = "top-right",
  panel_id = NULL,
  section_title = NULL,
  group_id = NULL
)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{id}{The ID for the custom control.}

\item{html}{The HTML content to add as a control.}

\item{position}{The position of the control on the map. Default is \code{"top-right"}.
Options include "top-left", "top-right", "bottom-left", "bottom-right".}

\item{panel_id}{ID of control panel to add to (optional).}

\item{section_title}{Section title when added to a control panel.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       The map or map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Add a custom HTML control to the map.
}
\examples{
if (interactive()) {
# Add to a map
map() |>
 add_custom_control(
   id = "custom_control",
   html = "<p>I am a custom control</p>"
)

# Add to a control panel
map() |>
 add_control_panel(panel_id = "my_panel", title = "Map Settings") |>
 add_custom_control(
   id = "custom_control_panel",
   html = "<p>I am a custom control in a panel</p>",
   panel_id = "my_panel",
   section_title = "Custom Control Section"
 )

# Add to a control panel inside a control group
map() |>
 add_control_panel(panel_id = "my_panel", title = "Map Settings") |>
 add_control_group(
   panel_id = "my_panel",
   group_id = "custom_controls",
   group_title = "Custom Controls"
   ) |>
 add_custom_control(
   id = "custom_control_panel",
   html = "<p>I am a custom control in a panel</p>",
   panel_id = "my_panel",
   group_id = "custom_controls"
 )
}
}
