% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{add_draw_control}
\alias{add_draw_control}
\title{Add a draw control to the map}
\usage{
add_draw_control(
  map,
  id = "draw_control",
  position = "top-right",
  modes = c("polygon", "trash"),
  active_colour = "#0FB3CE",
  inactive_colour = "#0FB3CE",
  mode_labels = list()
)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{position}{The position of the draw control on the map. Default is \code{"top-right"}.
Options are "top-left", "top-right", "bottom-left", "bottom-right".}

\item{modes}{A vector of modes to enable in the draw control.
Default is \code{c("polygon", "trash")}.
Options include "polygon", "trash", "line".}

\item{active_colour}{The colour for the drawn shapes. Default is \code{"#0FB3CE"}.}

\item{inactive_colour}{The colour for the inactive shapes. Default is \code{"#0FB3CE"}.}

\item{mode_labels}{A named list of labels for each mode.
For example, \code{list(polygon = "Draw Polygon", trash = "Delete Shape")}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{           The map or map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Add a draw control to the map
}
