% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{add_feature_server_source}
\alias{add_feature_server_source}
\title{Add a FeatureService source to the map.}
\usage{
add_feature_server_source(
  map,
  source_url,
  source_id,
  append_query_url = "/0/query?where=1=1&outFields=*&f=geojson"
)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{source_url}{The URL of the FeatureService source.}

\item{source_id}{The ID for the source.}

\item{append_query_url}{The query URL to append to the source URL. Default is
\code{"/0/query?where=1=1&outFields=*&f=geojson"}.}
}
\value{
The map or map proxy object for chaining.
}
\description{
Add a FeatureService source to the map.
}
\note{
By default the function appends a query URL to the provided \code{source_url} to retrieve all
features in GeoJSON format. If you need more control over the query parameters, you can
provide the full query URL directly in the \code{source_url} argument and set \code{append_query_url}
to an empty string to prevent appending the default query parameters.
}
\examples{
\dontrun{
service_url <- paste0(
 "https://services1.arcgis.com/VwarAUbcaX64Jhub/arcgis/rest/services/",
 "World_Exclusive_Economic_Zones_Boundaries/FeatureServer"
)

 map() |>
   add_feature_server_source(service_url, "eez") |>
   add_line_layer(id = "eez_lines", source = "eez")
}
}
