% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{add_feature_server_source}
\alias{add_feature_server_source}
\title{Add a FeatureServer source to the map.}
\usage{
add_feature_server_source(map, source_url, source_id)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{source_url}{The URL of the FeatureServer source.}

\item{source_id}{The ID for the source.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       The map or map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Add a FeatureServer source to the map.
}
