% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_fill_layer}
\alias{add_fill_layer}
\title{Add a fill layer to a map or map proxy.}
\usage{
add_fill_layer(map, ...)
}
\arguments{
\item{map}{The map object or map proxy to which the layer will be added.}

\item{...}{Additional arguments to include in the layer definition.
Default is an empty list.}
}
\value{
The updated map object with the fill layer added.
}
\description{
Add a fill layer to a map or map proxy.
}
\examples{
\dontrun{
# Load libraries
library(toro)
library(sf)

base_map <- map() |>
 add_feature_server_source(
   "https://services8.arcgis.com/AYGZtmUtpARUKBlB/arcgis/rest/services/Te_Reo_M\%C4\%81ori_Place_Names/FeatureServer/3/query?where=1=1&outFields=*&f=geojson",
   "maori_moana_data",
   append_query_url = FALSE
 ) |>
add_feature_server_source(
   "https://services8.arcgis.com/AYGZtmUtpARUKBlB/arcgis/rest/services/Te_Reo_M\%C4\%81ori_Place_Names/FeatureServer/1/query?where=1=1&outFields=*&f=geojson",
   "maori_region_data",
   append_query_url = FALSE
 )

base_map |>
 add_fill_layer(
   id = "moana_polygons",
   source = "maori_moana_data",
   hover_column = "name_mi",
   popup_column = "name"
 )

base_map |>
 add_fill_layer(
   id = "region_polygons",
   source = "maori_region_data",
   hover_column = "name_mi",
   popup_column = "name",
   paint = get_paint_options(
     "fill",
     options = list(
       colour = "#a3b18a",
       opacity = 0.3,
       outline_colour = "#588157"
     )
   )
 ) |>
 add_fill_layer(
   id = "moana_polygons",
   source = "maori_moana_data",
   hover_column = "name_mi",
   popup_column = "name",
   paint = get_paint_options(
     "fill",
     options = list(
       colour = get_column_group(
         "type",
         c(
           "lake" = "#a9d6e5",
           "reservoir" = "#89c2d9",
           "river" = "#61a5c2",
           "bay" = "#468faf",
           "sea" = "#2c7da0",
           "ocean" = "#2a6f97",
           "glacier" = "#014f86",
           "swamp" = "#01497c",
           "lagoon" = "#013a63"
         ),
         "#012a4a"
       )
     )
   )
 )
}
}
