% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_fill_layer}
\alias{add_fill_layer}
\title{Add a fill layer to a map or map proxy.}
\usage{
add_fill_layer(map, ...)
}
\arguments{
\item{map}{The map object or map proxy to which the layer will be added.}
}
\value{
The updated map object with the fill layer added.
}
\description{
Add a fill layer to a map or map proxy.
}
