% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{add_image}
\alias{add_image}
\title{Add an image source to the map.}
\usage{
add_image(map, image_id, image_url)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{image_id}{The ID of the image source.}

\item{image_url}{The URL of the image to add.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       The map or map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Add an image source to the map.
}
