% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_lat_lng_grid}
\alias{add_lat_lng_grid}
\title{Add a grid of latitude and longitude lines to the map.}
\usage{
add_lat_lng_grid(map, grid_colour = "#000000")
}
\arguments{
\item{map}{The map or map proxy object.}

\item{grid_colour}{The colour of the grid lines. Default is \code{"#000000"}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       The map or map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Add a grid of latitude and longitude lines to the map.
}
