% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_layer}
\alias{add_layer}
\title{Utilities for the map related to layers.}
\usage{
add_layer(
  map,
  id,
  type = "fill",
  source,
  paint = NULL,
  layout = NULL,
  popup_column = NULL,
  hover_column = NULL,
  can_cluster = FALSE,
  under_id = NULL,
  filter = NULL
)
}
\arguments{
\item{map}{The map object or map proxy to which the layer will be added.}

\item{id}{A unique identifier for the layer.}

\item{type}{The type of layer to add (e.g., "fill", "circle", "line").}

\item{source}{The data source for the layer, if not a GeoJSON, it will be converted.}

\item{paint}{A list of paint options for styling the layer.}

\item{layout}{A list of layout options for the layer.}

\item{popup_column}{The column name to use for popups. Default is \code{NULL}.}

\item{hover_column}{The column name to use for hover effects. Default is \code{NULL}.}

\item{can_cluster}{Whether the layer can be clustered. Default is \code{FALSE}.}

\item{under_id}{The ID of another layer to place this layer under. Default is \code{NULL}.}

\item{filter}{A filter expression to apply to the layer. Default is \code{NULL}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         The updated map object with the new layer added.
}\if{html}{\out{</div>}}
}
\description{
Functions:
\itemize{
\item \code{add_layer}: Add a layer to the map.
\item \code{add_fill_layer}: Add a fill layer to the map.
\item \code{add_circle_layer}: Add a circle layer to the map.
\item \code{add_line_layer}: Add a line layer to the map.
\item \code{add_symbol_layer}: Add a symbol layer to the map.
\item \code{add_lat_lng_grid}: Add a grid of latitude and longitude lines to the map.
\item \code{toggle_lat_lng_grid}: Show or hide the latitude and longitude grid.
\item \code{show_layer}: Show a previously hidden layer.
\item \code{hide_layer}: Hide a layer from the map.
\item \code{set_tile_layer}: Set the tile layer for the map.
\item \code{toggle_clustering}: Toggle clustering for a layer.
\item \code{set_paint_property}: Set a paint property for a layer.
\item \code{set_layout_property}: Set a layout property for a layer.
\item \code{get_tile_options}: Get available tile layer options.
Add a layer to a map or map proxy.
}
}
