% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_line_layer}
\alias{add_line_layer}
\title{Add a line layer to a map or map proxy.}
\usage{
add_line_layer(map, ...)
}
\arguments{
\item{map}{The map object or map proxy to which the layer will be added.}

\item{...}{Additional arguments to include in the layer definition.
Default is an empty list.}
}
\value{
The updated map object with the line layer added.
}
\description{
Add a line layer to a map or map proxy.
}
\examples{
\dontrun{
# Load libraries
library(toro)
library(sf)

base_map <- map() |>
 add_feature_server_source(
   "https://services8.arcgis.com/AYGZtmUtpARUKBlB/arcgis/rest/services/Te_Reo_M\%C4\%81ori_Place_Names/FeatureServer/4/query?where=1=1&outFields=*&f=geojson",
   "maori_awa_data",
   append_query_url = FALSE
 )

base_map |>
 add_line_layer(
   id = "awa_lines",
   source = "maori_awa_data",
   hover_column = "name_mi",
   popup_column = "name"
 )

base_map |>
 add_line_layer(
   id = "awa_lines",
   source = "maori_awa_data",
   hover_column = "name_mi",
   popup_column = "name",
   paint = get_paint_options(
     "line",
     options = list(
       colour = get_column_group(
         "type",
         c("river" = "#014f86", "stream" = "#61a5c2"),
         "#a9d6e5"
       ),
       line_width = get_column_group("type", c("river" = 3), 1),
       opacity = 0.8
     )
   )
 )
}
}
