% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_line_layer}
\alias{add_line_layer}
\title{Add a line layer to a map or map proxy.}
\usage{
add_line_layer(map, ...)
}
\arguments{
\item{map}{The map object or map proxy to which the layer will be added.}
}
\value{
The updated map object with the line layer added.
}
\description{
Add a line layer to a map or map proxy.
}
