% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{add_route}
\alias{add_route}
\title{Add a route to a toro map}
\usage{
add_route(map, route_id, points, settings = list())
}
\arguments{
\item{map}{A toro map object or a map proxy object}

\item{route_id}{A unique identifier for the route}

\item{points}{A sf object containing the points of the route}

\item{settings}{A list of settings for the route (e.g., color, weight)}
}
\value{
The updated map object
}
\description{
Add a route to a toro map
}
\examples{
\dontrun{
library(shiny)
library(toro)
library(sf)

line_data <- sf::st_sf(
  id = 1,
  geometry = sf::st_sfc(
    sf::st_linestring(
      cbind(c(172.2041, 163.9383), c(-32.56960, -46.43999))
    ),
    crs = 4326
  )
)

ui <- fluidPage(
 tagList(
   mapOutput("map"),
   actionButton("play_route", "Play Route Animation"),
   actionButton("pause_route", "Pause Route Animation"),
   actionButton("remove_route", "Remove Route")
 )
)
server <- function(input, output, session) {
 output$map <- renderMap({
   map() |>
     add_route(route_id = "route_line", points = line_data)
 })

 observe({
   req(input$map_loaded)
   mapProxy("map") |>
     play_route(route_id = "route_line")
 }) |>
   bindEvent(input$play_route)

  observe({
   req(input$map_loaded)
   mapProxy("map") |>
     pause_route(route_id = "route_line")
 }) |>
   bindEvent(input$pause_route)

  observe({
   req(input$map_loaded)
   mapProxy("map") |>
     remove_route(route_id = "route_line")
 }) |>
   bindEvent(input$remove_route)
}
}
}
