% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{add_route}
\alias{add_route}
\title{Add a route to a toro map.}
\usage{
add_route(map, route_id, points, settings = list())
}
\arguments{
\item{map}{A toro map object or a map proxy object.}

\item{route_id}{A unique identifier for the route.}

\item{points}{A sf object containing the points of the route.}

\item{settings}{A list of settings for the route (e.g., color, weight).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     The updated map object.
}\if{html}{\out{</div>}}
}
\description{
Add a route to a toro map.
}
