% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{add_source}
\alias{add_source}
\title{Utilities for the map related to map sources.}
\usage{
add_source(map, source_id, data, type = "geojson", cluster = FALSE)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{source_id}{The ID for the source.}

\item{data}{The data for the source, typically in GeoJSON format.}

\item{type}{The type of the source. Default is \code{"geojson"}.
Other options include \code{"vector"} or \code{"raster"}.}

\item{cluster}{Whether to enable clustering for this source. Default is \code{FALSE}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     The map or map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Functions:
\itemize{
\item \code{add_source}: Add a source to the map.
\item \code{add_feature_server_source}: Add a FeatureServer source to the map.
\item \code{add_image}: Add an image source to the map.
\item \code{set_source_data}: Set data for a source on the map.
Add a source to the map.
}
}
