% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_symbol_layer}
\alias{add_symbol_layer}
\title{Add a symbol layer to a map or map proxy.
This layer is typically used for icons or text labels.}
\usage{
add_symbol_layer(map, ...)
}
\arguments{
\item{map}{The map object or map proxy to which the layer will be added.}
}
\value{
The updated map object with the symbol layer added.
}
\description{
Add a symbol layer to a map or map proxy.
This layer is typically used for icons or text labels.
}
