% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_text_layer}
\alias{add_text_layer}
\title{Add a text layer to a map or map proxy}
\usage{
add_text_layer(map, ...)
}
\arguments{
\item{map}{The map object or map proxy to which the layer will be added}

\item{...}{Additional arguments to include in the layer definition.
\itemize{
\item clusterOptions: A list of options for clustering, if \code{can_cluster} is \code{TRUE}.
See the \href{https://epi-interactive-ltd.github.io/toro/articles/layers.html}{cluster vignette}
for details on available options
}}
}
\value{
The updated map object with the text layer added
}
\description{
This layer is typically used for text labels.
}
\examples{
\dontrun{
# Load libraries
library(toro)
library(sf)

# Prepare data
data(quakes)
quakes_data <- quakes |>
 st_as_sf(coords = c("long", "lat"), crs = 4326)

# Create map and add fill layer
map() |>
 add_text_layer(
   id = "test_layer",
   source = quakes_data
 )
}
}
