% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_control.R
\name{delete_drawn_shape}
\alias{delete_drawn_shape}
\title{Delete a drawn shape from the map.}
\usage{
delete_drawn_shape(proxy, shape_id)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{shape_id}{The ID of the shape to delete.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     The map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
The ID of the shape is provided by the draw control when a shape is created.
}
