% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layer_options.R
\name{get_layout_options}
\alias{get_layout_options}
\title{Get layout options for a specific layer type}
\usage{
get_layout_options(layer_type, options = list())
}
\arguments{
\item{layer_type}{A string indicating the type of layer (e.g., "fill", "circle", "line").}

\item{options}{A list of additional options to customize the layout properties.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A list of layout options suitable for the specified layer type.
}\if{html}{\out{</div>}}
}
\description{
This function returns a list of layout options based on the layer type and any additional
options provided.
}
\examples{
get_layout_options("line", list(line_cap = "butt", line_join = "bevel"))
get_layout_options("symbol", list(icon_image = "yellow_pin", icon_size = 1.5))

}
\seealso{
\code{\link{get_column}}, \code{\link{get_column_group}},
\code{\link{get_column_step_colours}}
}
