% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layer_options.R
\name{get_paint_options}
\alias{get_paint_options}
\title{Get paint options for a specific layer type}
\usage{
get_paint_options(layer_type, options = list())
}
\arguments{
\item{layer_type}{A string indicating the type of layer (e.g., "fill", "circle", "line").}

\item{options}{A list of additional options to customize the paint properties.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A list of paint options suitable for the specified layer type.
}\if{html}{\out{</div>}}
}
\description{
This function returns a list of paint options based on the layer type and any
additional options provided.
}
\examples{
get_paint_options("line", list(colour = "blue", opacity = 0.8, line_width = 2))
get_paint_options("circle", list(colour = "red", circle_radius = 10, outline_colour = "black"))
# Use with get_column for data-driven styling:
get_paint_options("fill", list(colour = get_column("color"), opacity = get_column("opacity")))
get_paint_options("fill", list(
   colour = get_column_group("group", c("A" = "green", "B" = "blue"))
))
get_paint_options("fill", list(
   opacity = get_column_step_colours("percent", c(25, 75), c("red", "orange", "yellow"))
))

}
\seealso{
\code{\link{get_column}}, \code{\link{get_column_group}},
\code{\link{get_column_step_colours}}
}
