% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{hide_layer}
\alias{hide_layer}
\title{Hide a layer from the map.}
\usage{
hide_layer(proxy, layer_id)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{layer_id}{The ID of the layer to hide.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     The map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Hide a layer from the map.
}
\note{
This does not remove the layer, it only hides it from view.
}
