% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{remove_control}
\alias{remove_control}
\title{Remove a control from the map.}
\usage{
remove_control(proxy, control_id)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{control_id}{The ID of the control to remove.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       The map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Remove a control from the map.
}
