% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{remove_control_group}
\alias{remove_control_group}
\title{Remove a control group from a control panel}
\usage{
remove_control_group(proxy, panel_id, group_id)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{panel_id}{The ID of the control panel.}

\item{group_id}{The ID of the control group to remove.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       The map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Remove a control group from a control panel
}
