% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{remove_custom_control}
\alias{remove_custom_control}
\title{Remove a custom control from the map.}
\usage{
remove_custom_control(proxy, control_id, panel_id = NULL)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{control_id}{The ID of the custom control to remove.}

\item{panel_id}{Optional. If provided, removes the custom control from the specified control panel.
If NULL, removes the standalone custom control.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       The map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Remove a custom control from the map.
}
\examples{
if (interactive()) {
# Add to a map
map() |>
 add_custom_control(
   id = "custom_control",
   html = "<p>I am a custom control</p>"
)
# In an observer
mapProxy("map") |>
 remove_custom_control("custom_control")

# Add to a control panel
map() |>
 add_control_panel(panel_id = "my_panel", title = "Map Settings") |>
 add_custom_control(
   id = "custom_control_panel",
   html = "<p>I am a custom control in a panel</p>",
   panel_id = "my_panel"
 )
# In an observer
mapProxy("map") |>
 remove_custom_control("custom_control", panel_id = "my_panel")
}
}
