% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{remove_custom_control}
\alias{remove_custom_control}
\title{Remove a custom control from the map.}
\usage{
remove_custom_control(proxy, control_id, panel_id = NULL)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{control_id}{The ID of the custom control to remove.}

\item{panel_id}{Optional. If provided, removes the custom control from the specified control panel.
If NULL, removes the standalone custom control.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       The map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Remove a custom control from the map.
}
