% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_control.R
\name{remove_draw_control}
\alias{remove_draw_control}
\title{Remove the draw control from the map}
\usage{
remove_draw_control(proxy, panel_id = NULL)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{panel_id}{Optional. If provided, removes the draw control from the specified control
panel. If NULL, removes the standalone draw control.}
}
\value{
The map proxy object for chaining.
}
\description{
Remove the draw control from the map
}
\examples{
if(interactive()){
library(shiny)
library(toro)

ui <- fluidPage(
 tagList(
   mapOutput("map"),
   actionButton("remove_draw_control", "Remove draw control")
 )
)
server <- function(input, output, session) {
 output$map <- renderMap({
   map() |>
    add_draw_control()
 })

 observe({
   req(input$map_loaded)
   mapProxy("map") |>
     remove_draw_control()
 }) |>
   bindEvent(input$remove_draw_control)
}
}
}
