% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_control.R
\name{remove_draw_control}
\alias{remove_draw_control}
\title{Remove the draw control from the map.}
\usage{
remove_draw_control(proxy, panel_id = NULL)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{panel_id}{Optional. If provided, removes the draw control from the specified control panel.
If NULL, removes the standalone draw control.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     The map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Remove the draw control from the map.
}
