% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{set_bounds}
\alias{set_bounds}
\title{Set the map bounds.}
\usage{
set_bounds(map, bounds, padding = 50, maxZoom = map$maxZoom)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{bounds}{A list of two lists, each containing longitude and latitude pairs.
For example: \code{list(list(-79, 43), list(-73, 45))}.}

\item{padding}{The padding around the bounds in pixels. Default is 50.}

\item{maxZoom}{The maximum zoom level to set. Default is the object's \code{maxZoom}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   The map or map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Set the map bounds.
}
