% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{set_layout_property}
\alias{set_layout_property}
\title{Set a layout property for a layer on the map.}
\usage{
set_layout_property(proxy, layer_id, property_name, value)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{layer_id}{The ID of the layer to update.}

\item{property_name}{The name of the layout property to set.}

\item{value}{The value to set for the layout property.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         The map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Set a layout property for a layer on the map.
}
