% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{set_tile_layer}
\alias{set_tile_layer}
\title{Set the tile layer for the map.}
\usage{
set_tile_layer(map, tiles)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{tiles}{A character vector of tile layer names.
Options include "natgeo", "satellite", "topo", "terrain",
"streets", "shaded", "light-grey".}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ The map or map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Set the tile layer for the map.
}
