% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{set_zoom}
\alias{set_zoom}
\title{Utilites for the map related to the view.}
\usage{
set_zoom(map, zoom)
}
\arguments{
\item{map}{The map or map proxy object.}

\item{zoom}{The zoom level to set. Default is 2.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ The map or map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Functions:
\itemize{
\item \code{set_zoom}: Set the map zoom level.
\item \code{set_bounds}: Set the map bounds.
Set the map zoom level.
}
}
