% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{show_layer}
\alias{show_layer}
\title{Show a previously hidden layer on the map.}
\usage{
show_layer(proxy, layer_id)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{layer_id}{The ID of the layer to show.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     The map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Show a previously hidden layer on the map.
}
