% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{toggle_clustering}
\alias{toggle_clustering}
\title{Toggle clustering for a layer on the map.}
\usage{
toggle_clustering(proxy, layer_id, cluster = FALSE)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{layer_id}{The ID of the layer to toggle clustering for.}

\item{cluster}{Whether to enable clustering. Default is \code{FALSE}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     The map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Toggle clustering for a layer on the map.
}
