% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{toggle_control}
\alias{toggle_control}
\title{Utilities for the map related to controls.}
\usage{
toggle_control(proxy, control_id, show = TRUE)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{control_id}{The ID of the control to toggle.}

\item{show}{Logical indicating whether to show or hide the control. Default is \code{TRUE}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       The map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Functions:
\itemize{
\item toggle_control:                Toggle the visibility of a control on the map.
\item remove_control:                Remove a control from the map.
\item add_custom_control:            Add a custom HTML control to the map.
\item remove_custom_control:         Remove a custom control from the map.
\item add_cursor_coords_control:     Add a cursor coordinates control to the map.
\item remove_cursor_coords_control:  Remove the cursor coordinates control from the map.
Toggle the visibility of a control on the map.
}
}
\examples{
if (interactive()) {
# In a Shiny app:
output$map <- renderMap({
 map() |>
   add_zoom_control() |>
   add_custom_control(
     id = "custom_control",
     html = "<p>I am a custom control</p>"
   )
})

# In an observer:
mapProxy("map") |>
 toggle_control("zoom_control", show = FALSE) |>
 toggle_control("custom_control", show = FALSE)
}
}
