% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{toggle_lat_lng_grid}
\alias{toggle_lat_lng_grid}
\title{Show/hide the latitude and longitude grid on the map.}
\usage{
toggle_lat_lng_grid(proxy, show = TRUE)
}
\arguments{
\item{proxy}{The map proxy object created by \code{mapProxy()}.}

\item{show}{Logical indicating whether to show or hide the grid. Default is \code{TRUE}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ The map proxy object for chaining.
}\if{html}{\out{</div>}}
}
\description{
Show/hide the latitude and longitude grid on the map.
}
